Allow to run capture_data.py without .env
capture_data.py has a dict with the server configuration and hence does
not need the corresponding env variables. However, the constants.py
checked the existence of these env vars unconditionally. Resulting in an
error message that the env vars are missing when running
capture_data.py.

This commit fixes this behaviour and will only check the existence of
env vars when they're actually required.